refactor: move ObjectObjectItem to its own file
